+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 <!-- RETAIN THIS COMMENT. An entry template for a new version is automatically added each time `Invoke-psake version` is called. Fill in changes afterwards. -->
 
+* **v1.3.2** (2021-04-16):
+  * [enhancement] `dbea`'s diagnostic output now uses `«` and `»` to enclose arguments, for improved visualization.
+  * [fix] PowerShell 7.2.0-preview.5 introduced a new experimental feature named `PSNativeCommandArgumentPassing`, aimed at solving the problem that `ie` solves in PowerShell itself.
+    However, it currently lacks vital accommodations for CLIs on Windows (see [GitHub issue #15143](https://github.com/PowerShell/PowerShell/issues/15143)).
+    This module now also checks for these accommodations and, if they're found missing, deactivates the feature in the module's scope by setting `$PSNativeCommandArgumentPassing` to `'Legacy'`, meaning that this module's workarounds are still being applied on top of the old, broken behavior, so as to provide all necessary accommodations.
+  * [fix] Partial double-quoting for `msiexec`-style applications for arguments such as `FOO="bar none"` now works again for batch-file calls (was accidentally broken when exit-code reporting was made reliable).
+
 * **v1.3.1** (2021-04-08):
-  * [enhancement] A `ins` invocation that passes separate arguments to the `-CommandLine` argument string now triggers a warning if these arguments aren't being referenced (such as with `$1` / `%1` on Unix / Windows). This alerts the use to accidental invocation of what should be `ins 'echo foo'` as `ins echo foo`.
+  * [enhancement] An `ins` invocation that passes separate arguments to the `-CommandLine` argument string now triggers a warning if these arguments aren't being referenced (such as with `$1` / `%1` on Unix / Windows). This alerts the user to accidental invocation of what should be `ins 'echo foo'` as `ins echo foo`.
 
 * **v1.3** (2021-04-06):
   * [enhancement] Reliable exit-code reporting for batch-file calls is now built into `ie`, via `cmd /c "<batch-file> ... & exit"`, courtesy of [this Stack Overflow post](https://stackoverflow.com/q/66975883/45375).